packages/iris-agent/pyproject.toml
[tool.poetry]
name = "iris-agent"
version = "0.1.0"
description = "The AI consciousness of the Phoenix ecosystem."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = "^0.27"
pydantic = "^2.6"
httpx = "^0.26"
google-generativeai = "^0.4.1"
phoenix_event_bridge = {path = "../phoenix_event_bridge", develop = true}
# Sécurité et authentification
pyjwt = "^2.8.0"
bcrypt = "^4.1.0"
# Monitoring et logging
slowapi = "^0.1.9"
# Base de données (si nécessaire pour l'auth)
supabase = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
